+++ src/layer/audio.ts
@@ -1,6 +1,6 @@
 // TODO: rename to something more consistent with the naming convention of Visual and VisualSourceMixin
 
-import { Base, BaseOptions } from './base'
+import { BaseAudio, BaseAudioOptions } from './base-audio'
 import { AudioSourceMixin, AudioSourceOptions } from './audio-source'
 
 type AudioOptions = AudioSourceOptions
@@ -8,7 +8,7 @@ type AudioOptions = AudioSourceOptions
 /**
  * @extends AudioSource
  */
-class Audio extends AudioSourceMixin<BaseOptions>(Base) {
+class Audio extends AudioSourceMixin<BaseAudioOptions>(BaseAudio) {
   /**
    * Creates an audio layer
    */