+++ src/layer/index.ts
@@ -5,6 +5,13 @@
 
 export * from './audio-source'
 export * from './audio'
+export {
+  BaseAudioMixin,
+  BaseAudioOptions
+} from './base-audio-mixin'
+export {
+  BaseAudio
+} from './base-audio'
 export * from './base'
 export * from './image'
 export * from './text'