+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- `etro.layer.BaseAudio` (which all audio layers subclass).
+
 ### Fixed
 - `Audio` and `Video` layers not detaching properly.